move /deployment to /yatai/deployment
