+++ bentoml/cli/aws_lambda.py
@@ -30,7 +30,7 @@ from bentoml.cli.deployment import (
     _print_deployment_info,
     _print_deployments_info,
 )
-from bentoml.deployment.store import ALL_NAMESPACE_TAG
+from bentoml.yatai.deployment.store import ALL_NAMESPACE_TAG
 from bentoml.exceptions import BentoMLException
 from bentoml.yatai.proto import status_pb2
 from bentoml.yatai.proto.deployment_pb2 import DeploymentSpec
