+++ bentoml/yatai/client/deployment_api.py
@@ -19,7 +19,7 @@ import logging
 import time
 
 from bentoml.utils import status_pb_to_error_code_and_message
-from bentoml.deployment.store import ALL_NAMESPACE_TAG
+from bentoml.yatai.deployment.store import ALL_NAMESPACE_TAG
 from bentoml.yatai.proto.deployment_pb2 import (
     ApplyDeploymentRequest,
     DescribeDeploymentRequest,
