+++ bentoml/yatai/deployment/sagemaker/sagemaker_wsgi.py
@@ -15,7 +15,7 @@
 import os
 
 from bentoml.saved_bundle import load
-from bentoml.deployment.sagemaker.model_server import BentomlSagemakerServer
+from bentoml.yatai.deployment.sagemaker.model_server import BentomlSagemakerServer
 
 api_name = os.environ.get('API_NAME', None)
 model_service = load('/opt/program')
